mlestar/requirements.txt
deepagents>=0.4.12
langchain-core>=0.3.0
langgraph>=0.4.0
langgraph-checkpoint>=2.0.0
pydantic>=2.0.0
pyyaml>=6.0
requests>=2.31.0
